fix issues based on comments from Jeffrey
